+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 _Please, add the latest build info on top of the list; use Version::MAJOR only after gold release; keep semantic versioning in line with framework's_
 
+- **0.5.12** [Steve A.] added LastSeasonId view; DB vers. 2.00.0
+- **0.5.11** [Steve A.] added #by_swimmer to AbstractResult; specs fixes & update
 - **0.5.10** [Steve A.] added #active? to Users; refactored all to_hash methods using ancestor AbstractRecord class; specs & bundle update
 - **0.5.05** [Steve A.] added to_hash methods as middlemen in between to_json output
 - **0.5.03** [Steve A.] slight change in ImportQueueDecorator so that #chrono_delta_label can be called also on master chrono rows